+++ packages/tswebextension/src/lib/mv2/background/request/request-blocking-api.ts
@@ -34,6 +34,10 @@ type RequestParams = Pick<
 export type GetBlockingResponseParams = RequestParams & {
     rule: NetworkRule | null;
     popupRule: NetworkRule | null;
+    /**
+     * Indicates whether the request is a prerender request.
+     */
+    isPrerenderRequest?: boolean;
 };
 
 /**
@@ -142,6 +146,7 @@ export class RequestBlockingApi {
             requestUrl,
             requestId,
             referrerUrl,
+            isPrerenderRequest,
         } = data;
 
         if (!rule) {
@@ -201,6 +206,7 @@ export class RequestBlockingApi {
                     referrerUrl,
                     rule,
                     tabId,
+                    isPrerenderRequest,
                 });
             }