+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased <!-- release/browser-extension-v5.4 -->
+
+### Added
+
+- Correct handle prerender speculative requests in Chromium-based browsers.
+
 ## [4.0.0] - TBD
 
 ### Added